pyproject.toml
[tool.poetry]
name = "business-mgmt"
version = "0.1.0"
description = ""
authors = ["holyhloy <tlegchilin@yandex.ru>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = { extras = ["standard"], version = ">=0.34.2,<0.35.0" }
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
asyncpg = ">=0.30.0,<0.31.0"
pydantic = ">=2.11.4,<3.0.0"
pydantic-settings = ">=2.9.1,<3.0.0"
fastapi-users = { extras = ["sqlalchemy"], version = ">=14.0.1,<15.0.0" }
redis = ">=6.1.0,<7.0.0"
celery = ">=5.5.2,<6.0.0"
httpx = ">=0.28.1,<0.29.0"
jinja2 = ">=3.1.6,<4.0.0"
sqladmin = ">=0.20.1,<0.21.0"
itsdangerous = ">=2.2.0,<3.0.0"
apscheduler = ">=3.11.0,<4.0.0"
fastapi-cache2 = ">=0.2.2,<0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
pytest-html = "^4.1.1"
pytest-env = "^1.1.5"
pytest-mock = "^3.14.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[tool.pytest.ini_options]
env = [
    "DB_NAME=business_management_test",
    "MODE=TEST"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
